requirements-deepseek.txt
# Dependencies untuk DeepSeek V3 API Integration
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0

# OpenAI SDK untuk DeepSeek API (kompatibel)
openai>=1.0.0

# Environment management
python-dotenv>=0.19.0

# Machine learning utilities
scikit-learn>=1.1.0

# Logging
loguru>=0.6.0

# Progress bars
tqdm>=4.64.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Data validation
jsonschema>=4.0.0

# Path utilities
pathlib2>=2.3.0

# Type hints
typing-extensions>=4.0.0

# Retry mechanism (optional)
retrying>=1.3.3